refined some text and recombination process

+++ src/main/scala/ch/uzh/ifi/pdeboer/pplib/hcomp/mturk/MTurkProcessor.scala
@@ -41,7 +41,7 @@ class MTurkManager(val service: MTurkService, val query: HCompQuery, val propert
 		val mtQuery = MTQuery.convert(query)
 
 		val ONE_DAY: Int = 60 * 60 * 24
-		val dollars: Double = properties.paymentCents / 100
+		val dollars: Double = properties.paymentCents.toDouble / 100d
 		val hitTypeID = service.RegisterHITType(query.title, query.question, Price(dollars.toString), ONE_DAY, Seq.empty[String], ONE_DAY, Seq.empty[QualificationRequirement])
 		hit = service.CreateHIT(hitTypeID, new Question(mtQuery.xml), ONE_DAY, 1).HITId
 		hit
